update:sms

+++ src/bedis/SmsString.php
@@ -4,7 +4,7 @@ namespace think\bit\bedis;
 
 use think\bit\common\Bedis;
 
-class SmsString extends Bedis
+final class SmsString extends Bedis
 {
     protected $key = 'sms:';
 
@@ -18,7 +18,12 @@ class SmsString extends Bedis
     function factory($phone, $code, $timeout = 60)
     {
         try {
-            return $this->redis->set($this->key . $phone, $code, $timeout);
+            $data = msgpack_pack([
+                'code' => $code,
+                'publish_time' => time(),
+                'timeout' => $timeout
+            ]);
+            return $this->redis->set($this->key . $phone, $data, $timeout);
         } catch (\Exception $e) {
             return false;
         }
@@ -28,16 +33,43 @@ class SmsString extends Bedis
      * 验证手机验证码
      * @param string $phone 手机号
      * @param string $code 验证码
-     * @param boolean $once 验证一次有效
+     * @param boolean $once 验证一次有效,验证完成即删除
      * @return bool
      */
     function check($phone, $code, $once = false)
     {
-        $result = ($code === $this->redis->get($this->key . $phone));
-        if ($once && $result) {
-            $this->redis->delete($this->key . $phone);
+        try {
+            if (!$this->redis->exists($this->key . $phone)) return false;
+            $data = msgpack_unpack($this->redis->get($this->key . $phone));
+
+            $result = ($code === $data['code']);
+            if ($once && $result) {
+                $this->redis->delete($this->key . $phone);
+            }
+
+            return $result;
+        } catch (\Exception $e) {
+            return false;
         }
-        return $result;
     }
 
+    /**
+     * 获取验证时间
+     * @param string $phone 手机号
+     * @return array|bool
+     */
+    function time($phone)
+    {
+        try {
+            if (!$this->redis->exists($this->key . $phone)) return false;
+            $data = msgpack_unpack($this->redis->get($this->key . $phone));
+
+            return [
+                'publish_time' => $data['publish_time'],
+                'timeout' => $data['timeout']
+            ];
+        } catch (\Exception $e) {
+            return false;
+        }
+    }
 }
